cms/requirements/base.txt
djangorestframework >= 3.9.4
django-cors-headers >= 3.0.1
google-api-python-client == 1.7.9
google-auth-httplib2 == 0.0.3
google-auth-oauthlib == 0.4.0